Initial refactoring of the upload document controllers
